pkg/report: ignore transient crashes
There are cases when a crash, even though it didn't occur during boot
and image testing phases, still should not be counted during bisection.
Otherwise we risk diverting the whole process in the wrong direction.

One of such problem classes is getting SYZFATAL errors when we are not
bisecting a SYZFATAL crash.