pkg/repro: tolerate two consequential run errors
Retrying once has greatly reduced the number of "failed to copy prog to
VM" errors, but they still periodically pop up. The underlying problem
is still not 100% known.

Supposedly, if a booted VM with an instrumented kernel has to wait too long,
it can just hang or crash by itself. At least on some problematic
revisions.

Investigation wouldbe quite time-consuming -- we need to do a
complicated refactoring in order to also capture serial output for
Copy() failures. So far it does not seem to be totally worth it.

Let's do 3 runOnInstance() attempts. If the problem still persists,
there's no point in doing more runs -- we'd have to determine the
exact root cause.