pkg/vcs: circumvent ref conflicts during fetch
If the remote repo's tags conflict with the local ones, `git fetch` may
fail with `error: cannot lock ref %: % exists`.

It seems that the only way to restore the repository in this case is to
do `git fetch` with `--prune --prune-tags`. Since this also forces the
repository to forget all tags not present in the remote, let's do it
only when we got the error message.