compiler: support const as int first argument
This commit adds support for the following syntax:

    int8[constant]

as an equivalent to:

    const[constant, int8]

The goal is to have a unified const/flags definition that we can use in
templates. For example:

    type template[CLASS, ...] {
        class int8:3[CLASS]
        // ...
    }

    type singleClassType template[SINGLE_CONST]
    type subClassType template[abc_class_flags]

In this example, the CLASS template field can be either a constant or a
flag. This is especially useful when defining both a generic instance of
the template as well as specialized instances (ex. bpf_alu_ops and
bpf_add_op).

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>