pkg/repro: fix potential deadlock
If an instance fails to boot 3 times in a row,
we drop it on the floor. If we drop all instances this way,
the repro process deadlocks infinitly.
Retry infinitly instead. There is nothing else good we can do
in this case. If the instances becomes alive again, repro will resume.