pkg/vcs: support fetches by a short git hash
The approach we used works perfectly for all commits, but they must be
referenced by the full 40 character hash. In almost all cases, users
would prefer to use the shorter one.

If the commit hash is not 40 characters long, fetch the whole git tree.

The only unsupported scenario is fetching a commit that is referenced by
custom refs/* by its short hash. It's unlikely there's anything we can
do here.